create list and update partenaire is done

+++ src/main/java/com/transfert/isi/transfert/services/PartenaireService.java
@@ -46,7 +46,7 @@ public interface PartenaireService {
      * @param id the id of the entity
      * @return the entity with all Eager
      */
-    Optional<Partenaire> findOneWithEager(Long id);
+    Optional<Partenaire> findOneWithEager(Integer id);
 
     /**
      * Delete the "id" Partenaire.
